rename SendingClock to EchoingClock
